# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance Go + Gin + PostgreSQL 17 CRM Lead Management backend.",
        "title": "Universal CRM - Lead Management API (Go)",
        "contact": {},
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/webhooks/lead": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Ingest lead from external webhooks using Bearer API Key",
                "parameters": [
                    {
                        "description": "Webhook Lead Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.WebhookLeadPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.WebhookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.WebhookLeadPayload": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "custom_fields": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_address": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "lead_source": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "requirement": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "handlers.WebhookResponse": {
            "type": "object",
            "properties": {
                "lead": {
                    "$ref": "#/definitions/models.LeadResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.CustomFieldValueItem": {
            "type": "object",
            "properties": {
                "field_key": {
                    "type": "string"
                },
                "field_label": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "models.LeadResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "custom_field_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CustomFieldValueItem"
                    }
                },
                "custom_fields": {
                    "type": "object",
                    "additionalProperties": true
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "follow_up_completed": {
                    "type": "boolean"
                },
                "follow_up_date": {
                    "type": "string"
                },
                "follow_up_notes": {
                    "type": "string"
                },
                "follow_up_type": {
                    "type": "string"
                },
                "last_activity_at": {
                    "type": "string"
                },
                "lead_age_days": {
                    "type": "integer"
                },
                "lost_reason": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes_count": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "owner_id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "requirement": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.UserSummary": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}